pyproject.toml
[tool.poetry]
name = "Flask_test_app"
version = "0.1.0"
description = ""
authors = ["ku4umid3e <kuchumid3e@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.2"
requests = "^2.27.1"
beautifulsoup4 = "^4.10.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
